implement modular based relative position embedding

+++ binary_model.py
@@ -77,7 +77,7 @@ def get_attn_pos(attn_mask, num_local=16):
     # pos_emb = pos_embedding(pos_ind, d_word_vec, wave_length=10000)
     pos_ind = torch.from_numpy(pos_ind).unsqueeze(0).expand(attn_shape + (2,))
     if attn_mask.is_cuda:
-        pos_ind = pos_ind.cuda()
+        pos_ind = pos_ind.cuda().float()
         print(type(pos_ind))
     return pos_ind
